% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cog_open}
\alias{cog_open}
\title{Open a source (local path, http(s)://, s3://, gs://, az://). Returns an
external pointer reused by \code{cog_meta()} / \code{cog_fetch_window()}.}
\usage{
cog_open(src)
}
\description{
Open a source (local path, http(s)://, s3://, gs://, az://). Returns an
external pointer reused by \code{cog_meta()} / \code{cog_fetch_window()}.
}
